+++ cron/data/index.ts
@@ -1,5 +1,9 @@
 import { blockchain, tmapApi } from "../../core/instance";
-import { averageInstance, cacheOrderInstance, locationInstance } from "../../maria/commands";
+import {
+  averageCostRepository,
+  deliveryPersonMatchedDateRepository,
+  locationRepository,
+} from "../../database/type-orm/instance";
 
 import { Combiner } from "./combiner";
 import { DataService } from "./data-service";
@@ -7,9 +11,9 @@ import { DB } from "./database";
 import { ExternalApi } from "./external-api";
 
 const database = new DB({
-  averageCostRepository: averageInstance,
-  deliveryPersonMatchedDateRepository: cacheOrderInstance,
-  locationInstance,
+  averageCostRepository: averageCostRepository,
+  deliveryPersonMatchedDateRepository: deliveryPersonMatchedDateRepository,
+  locationRepository: locationRepository,
 });
 const externalApi = new ExternalApi({ blockchain, tmapApi });
 const combiner = new Combiner();
