+++ database/type-orm/repository/impl/index.ts
@@ -1,4 +1,7 @@
 export * from "./average-cost.repository";
 export * from "./chat-room.repository";
 export * from "./delivery-person-matched-date-repository";
+export * from "./location";
+export * from "./order";
 export * from "./receiver.repository";
+export * from "./user";
