+++ database/type-orm/repository/impl/location/index.ts
@@ -1,0 +1,2 @@
+export * from "./location.repository";
+export * from "./location.repository.impl";
