+++ database/type-orm/repository/impl/order/order.repository.ts
@@ -1,7 +1,7 @@
 import { EntityManager } from "typeorm";
-import { TRANSPORTATION_ALLOW_VALUES } from "../../../../validator/schema/routes/order/custom-validator";
-import { OrderControllerRequestData } from "../../../../validator/schema/routes/order/order-controller-request-data";
-import { Order } from "../../entity";
+import { TRANSPORTATION_ALLOW_VALUES } from "../../../../../validator/schema/routes/order/custom-validator";
+import { OrderControllerRequestData } from "../../../../../validator/schema/routes/order/order-controller-request-data";
+import { Order } from "../../../entity";
 
 type BasicTransportationEntity = {
   transportation: Record<(typeof TRANSPORTATION_ALLOW_VALUES)[number], 1 | 0>;
