+++ database/type-orm/repository/impl/user/index.ts
@@ -1,0 +1,2 @@
+export * from "./user.repository";
+export * from "./user.repository.impl";
