+++ database/type-orm/repository/index.ts
@@ -1,3 +1,2 @@
 export * from "./abstract-repository";
 export * from "./impl";
-export * from "./order";
